Modularizes code by moving fireball challenge from main code to the challenges file
